New file added

+++ circumference.java
@@ -5,10 +5,10 @@ public class circumference {
     public static void main(String[] args) {
         System.out.print("Enter radius of circle=");
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        float n = sc.nextInt();
         // int circumference=2*n*(22/7);
-        int  PI =22/7 ;
-        int circumference = 2 * n * PI;
+        float PI =22/7 ;
+        float circumference = 2 * n * PI;
 
         // System.out.print("Circumference="+circumference);
         System.out.println(2 * n * PI);
